fix(blockchain-link):  types circular references
