+++ packages/blockchain-link/src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './common';
+export * from './params';
 
 export { Message } from './messages';
 export { Response } from './responses';
